requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
dnspython==2.6.1
psycopg2-binary==2.9.9
pymongo==4.7.2
sqlparse==0.5.0
